requirements.txt
asgiref==3.8.1
Django==5.2.4
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2